switch the various id private elements and getter functions to use int64 to simplify some db tasks

+++ actor/actor.go
@@ -45,7 +45,7 @@ type Actor interface {
 	ACLName() string
 	Authz() string
 	IsACLRole() bool
-	GetId() int
+	GetId() int64
 }
 
 // GetReqUser gets the actor making the request. If use-auth is not on, always
