+++ client/client.go
@@ -58,7 +58,7 @@ type Client struct {
 	Certificate string `json:"certificate"`
 	org         *organization.Organization
 	AuthzID     string `json:"authz_id"`
-	id int
+	id int64
 }
 
 // for gob encoding. Needed the json tags for flattening, but that's handled
@@ -734,6 +734,6 @@ func (c *Client) ACLName() string {
 	return c.GetName()
 }
 
-func (c *Client) GetId() int {
+func (c *Client) GetId() int64 {
 	return c.id
 }
